requirements.txt
astroid==2.12.13
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dill==0.3.6
Flask==2.2.2
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
mysqlclient==2.1.1
parso==0.8.3
platformdirs==2.5.4
PyJWT==2.6.0
pylint==2.15.8
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.44
tomlkit==0.11.6
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
